+++ packages/multichain-api-middleware/src/handlers/wallet-createSession.test.ts
@@ -4,11 +4,11 @@ import {
   type Caip25Authorization,
   type NormalizedScopesObject,
   KnownSessionProperties,
+  Caip25Errors,
 } from '@metamask/chain-agnostic-permission';
 import * as ChainAgnosticPermission from '@metamask/chain-agnostic-permission';
 import { MultichainNetwork } from '@metamask/multichain-transactions-controller';
 import { invalidParams } from '@metamask/permission-controller';
-import { JsonRpcError, rpcErrors } from '@metamask/rpc-errors';
 import type {
   Hex,
   Json,
@@ -17,14 +17,6 @@ import type {
 } from '@metamask/utils';
 
 import { walletCreateSession } from './wallet-createSession';
-
-jest.mock('@metamask/rpc-errors', () => ({
-  ...jest.requireActual('@metamask/rpc-errors'),
-  rpcErrors: {
-    invalidParams: jest.fn(),
-    internal: jest.fn(),
-  },
-}));
 
 jest.mock('@metamask/chain-agnostic-permission', () => ({
   ...jest.requireActual('@metamask/chain-agnostic-permission'),
@@ -179,7 +171,7 @@ describe('wallet_createSession', () => {
       },
     });
     expect(end).toHaveBeenCalledWith(
-      new JsonRpcError(5302, 'Invalid sessionProperties requested'),
+      Caip25Errors.invalidSessionPropertiesError(),
     );
   });
 
@@ -567,7 +559,7 @@ describe('wallet_createSession', () => {
       });
     await handler(baseRequest);
     expect(end).toHaveBeenCalledWith(
-      new JsonRpcError(5100, 'Requested scopes are not supported'),
+      Caip25Errors.requestedChainsNotSupportedError(),
     );
   });
 
@@ -888,8 +880,8 @@ describe('wallet_createSession', () => {
     });
   });
 
-  it('calls internal RPC error if approved CAIP-25 permission has no CAIP-25 caveat value', async () => {
-    const { handler, requestPermissionsForOrigin } = createMockedHandler();
+  it('calls end with error if approved CAIP-25 permission has no CAIP-25 caveat value', async () => {
+    const { handler, requestPermissionsForOrigin, end } = createMockedHandler();
     requestPermissionsForOrigin.mockReturnValue([
       {
         [Caip25EndowmentPermissionName]: {
@@ -910,7 +902,9 @@ describe('wallet_createSession', () => {
       },
     });
 
-    expect(rpcErrors.internal).toHaveBeenCalled();
+    expect(end).toHaveBeenCalledWith(
+      Caip25Errors.unknownErrorOrNoScopesAuthorized(),
+    );
   });
 
   describe('address case sensitivity', () => {
@@ -1290,7 +1284,7 @@ describe('wallet_createSession', () => {
       await handler(requestWithNoValidScopes);
 
       expect(end).toHaveBeenCalledWith(
-        new JsonRpcError(5100, 'Requested scopes are not supported'),
+        Caip25Errors.requestedChainsNotSupportedError(),
       );
     });
   });